+++ test/support/fixtures/v1_lead_sources.ex
@@ -1,0 +1,15 @@
+defmodule ExPipedrive.Fixtures.V1LeadSources do
+  @moduledoc false
+
+  def list_response do
+    %{
+      "success" => true,
+      "data" => [
+        %{"name" => "Manually created"},
+        %{"name" => "Deal"},
+        %{"name" => "Web forms"},
+        %{"name" => "API"}
+      ]
+    }
+  end
+end